services/wavestream/requirements.txt
yt-dlp~=2025.2.19
requests~=2.32.3
numpy~=2.2.3
Flask~=3.1.0
pillow~=11.1.0
#streamlink~=7.1.3
#hitherdither @ git+https://github.com/hbldh/hitherdither@0f3bbc4
